Improve computing speed of preprocessor
Since the complex normalizer is very slow, this patch fixes that.
According to stanza
(https://github.com/stanfordnlp/stanza/blob/master/README.md#batching-to-maximize-pipeline-speed)
"the best approach" to solve that is by batch documents together.
This what this patch is about.
However it also filters out of a Stack of Documents, Documents which had
been processed before in the given variant.
It also decreases the load of cpu alot and needs ~8GB of memory per process (with 2 workers).
Most importantly it boils down the excution time (with 2 workers) of the preprocessor
from 8h+ to 45 min regarding the trainings data.
This patch also drops out the ner processor (since we do not take
advantage out of it) and increases the batch size for lemmas as well as
for tokens.