Add Feedforward neural network model
This patch replaces the current model with a FFN Model. Therefore it
makes some auxilary definitions as well as it changes the property
propagation. (e.g. workers are now seperated for each phase)
Since we agreed that properties of a model are hard wired it also
removes the model parameter from the PropertiesMananger.