HookUp ServiceLocator Part II
In order to make the TextminingManager be able to use the
ServiceLocator, this patch redesigns the MLP/MLPManager and their test.
This terminates the coupling of test, training and validation of a model
as well as makes some minor bug fixes in terms of keras usage. The patch
makes also a clear seperation of a abstract Base Class (MLPBase) and
the MLP Interface.
The MLPManager acts now as a micro ServiceLocator or in this a case as a
MLPModelLocator.