HookUp ServiceLocator Part III
In order to prepare the integration of a proper evaluator, this patch
splits out the vectorizer of TextMiningManger.
However the Vectorizer has now 2 stages - feature extration and
selection. So far there is no feature selection method implemented, but
will be done after the Evalution is finished.
Also this patch makes some smoothment in order of the Name of
Arguments/Properties.